Fixed static variable bug in log_msg causing crashes on 1M testing.
A PyObject pointer was declared static, used and DECREFed in log_msg,
causing segfaults when running on >1 FEM with API tracing on. This was
due to multiple threads calling log_msg and there being a race condition
on use of that variable. Fixed by removing static but a simpler
implementation not passing via a PyObject has been used anyway.