Add redirectable logging to API extension.
This commit implements a simple FemLogger class for the API extension module,
replacing all std::cout messaging. FemLogger can be configured with an external
logging function: this is set up in the wrapper to call the Python logging module
functions, allowing the API module to integrate its logging with the enclosing
Python module and application.