Add frame counter reset command
This change adds a reset of the UDP core frame and packet counters to
the acquisition start logic. This is done for all modes except 24-bit C0
matrix read, where it is desirable that the frame counter continue. Also
reordered some of the startAcquisition code sequence for simpler
reading. Required FEM f/w version 119 or greater.