Switch FR to use frame header for frame counter.
With new v11A FEM firware, the frame counter in the UDP packet header
is reset at the start of each acquisition. This makes that number
a more reliable source of the current (sub)frame number than the
subframe trailer, which is not present in all modes and is reset
for each frame in e.g. DAC scans and 24-bit mode. This avoids the
complex logic associated with tracking and shadowing subframes.