Fix memory leak in API wrapper methods.
This fixes a per-call memory leak in the FEM API wrapper _get_XXX
functions, which build a list to return to the calling function.
Py_BuildValues() doesn't take ownership of the reference count, so it
is necessary to decouple that call from the return so that Py_DECREF
can be called on the list first.